requirements.txt
# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# HTTP client (modern replacement for requests)
httpx[http2]>=0.25.0

# LLM API
groq>=0.4.0

# Environment variables
python-dotenv>=1.0.0

# Retry logic for robustness
tenacity>=8.2.0

# Note: pytz removed - using built-in zoneinfo (Python 3.9+)
